chore: add ClientDebug method to test errors
- moved ClientDebug class to a new file to avoid circular dependency errors

- introduced ReportFatalErrorWithUnityPayloadAsync for easier use of async/await syntax
