+++ kernel/packages/unity-interface/dcl.ts
@@ -25,7 +25,8 @@ import { StoreContainer } from 'shared/store/rootTypes'
 import { ILandToLoadableParcelScene, ILandToLoadableParcelSceneUpdate } from 'shared/selectors'
 import { UnityParcelScene } from './UnityParcelScene'
 import { loginCompleted } from 'shared/ethereum/provider'
-import { ClientDebug, clientDebug, UnityInterface, unityInterface } from './UnityInterface'
+import { UnityInterface, unityInterface } from './UnityInterface'
+import { clientDebug, ClientDebug } from "./ClientDebug"
 import { BrowserInterface, browserInterface } from './BrowserInterface'
 import { UnityScene } from './UnityScene'
 import { ensureUiApis } from 'shared/world/uiSceneInitializer'
